:hammer: refactor : 조회수 초기화 및 날짜별 조회수 저장 로직 수정
