+++ src/main/java/hitit/post/PostRepository.java
@@ -1,10 +1,10 @@
-package hitit.post;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post,Long> {
-
-    Post findByPostId(Long postId);
-}
+//package hitit.post;
+//
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface PostRepository extends JpaRepository<Post,Long> {
+//
+//    Post findByPostId(Long postId);
+//}
